requirements.txt
asgiref==3.5.2
Django==4.1.1
django-crispy-forms==1.14.0
django-filter==22.1
install==1.3.5
Pillow==9.2.0
sqlparse==0.4.2
